Refactor: replace magic numbers with named constants and fix clippy warnings
Named Constants:
- Add OF_CLIENT_INTERFACE_ADDR (0xFFF1FFF0)
- Add HIGH_MEM_IO_BASE (0xFFFF0000)
- Add HIGH_MEM_IO_SIZE (0x10000)

Benefits:
- More maintainable - single source of truth
- Self-documenting - names explain purpose
- Easier to change - update once vs. multiple locations

Clippy Fixes:
- Remove unnecessary cast in AdbMouse::poll()
- Simplify DiagnosticDevice::read() match patterns
- Use match guards for cleaner code

All 70 tests pass

+++ crates/newton-devices/src/adb.rs
@@ -147,7 +147,7 @@ impl AdbDevice for AdbMouse {
     fn poll(&mut self) -> Option<AdbPacket> {
         // Return current mouse state
         let data = vec![
-            (self.buttons & 0x80) as u8,
+            self.buttons & 0x80,
             (self.x >> 8) as u8,
             (self.x & 0xFF) as u8,
             (self.y >> 8) as u8,
